# sstables/trie: translate all key cells in one go, not lazily