--- a/configure.py
+++ b/configure.py
@@ -618,6 +618,7 @@ perf_tests = set([
     'test/perf/perf_idl',
     'test/perf/perf_vint',
     'test/perf/perf_big_decimal',
+    'test/perf/perf_bti_key_translation',
     'test/perf/perf_sort_by_proximity',
 ])