--- a/test/perf/CMakeLists.txt
+++ b/test/perf/CMakeLists.txt
@@ -94,3 +94,7 @@ add_perf_test(perf_row_cache_reads)
 add_perf_test(perf_generic_server)
 add_perf_test(perf_s3_client)
 add_perf_test(perf_sort_by_proximity)
+add_perf_test(perf_bti_key_translation
+  LIBRARIES
+    dht
+    sstables)