@avikivity 2025-10-01
Wow, 5X performance, 6 IPC.